meteoblue_wolf/pyproject.toml
[project]
name = "meteoblue-wolf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "common",
    "geopandas>=1.0.1",
    "pandas>=2.2.3",
    "pycryptodome>=3.22.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[tool.uv.sources]
common = { git = "https://github.com/opendatabs/common", rev = "v0.1.0" }

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
